+++ src/components/TodoList/index.ts
@@ -1,0 +1,1 @@
+export { TodoList } from './TodoList';
